+++ packages/react-layout/src/website/demos/item-wizard/demo.ete.js
@@ -1,0 +1,7 @@
+/**
+  * Copyright 2020, Verizon Media
+  * Licensed under the terms of the MIT license. See LICENSE file in project root for terms.
+  */
+
+export default async function () {
+}