+++ packages/react-layout/src/website/demos/item-wizard/form/index.js
@@ -1,0 +1,14 @@
+import components from './components';
+import data from './data';
+import fields from './fields';
+
+export default {
+  model: {
+    id: 'simple',
+    data,
+    fields,
+  },
+  resources: {
+    components,
+  },
+};