+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Changes that have landed in master but are not yet released
 
+### React Layout
+
+* Add - wizard demo
+
 ## v1.0.6 (July 5, 2020)
 
 ### Form